Fixed camera automatic scrolling processing - Crushed playable character will be killed - Fixed black border at the left side fixed issues #145 and #80

+++ Engine/common_features/fader.cpp
@@ -22,55 +22,55 @@
 
 PGE_Fader::PGE_Fader()
 {
-    cur_ratio=0.0;
-    target_ratio=0.0;
-    fade_step=0.0;
-    fadeSpeed=0;
-    manual_ticks=0;
+    cur_ratio    = 0.0;
+    target_ratio = 0.0;
+    fade_step    = 0.0;
+    fadeSpeed    = 0;
+    manual_ticks = 0;
 }
 
 PGE_Fader::PGE_Fader(const PGE_Fader &f)
 {
-    cur_ratio      =f.cur_ratio;
-    target_ratio   =f.target_ratio;
-    fade_step      =f.fade_step;
-    fadeSpeed      =f.fadeSpeed;
-    manual_ticks   =f.manual_ticks;
+    cur_ratio    = f.cur_ratio;
+    target_ratio = f.target_ratio;
+    fade_step    = f.fade_step;
+    fadeSpeed    = f.fadeSpeed;
+    manual_ticks = f.manual_ticks;
 }
 
 PGE_Fader::~PGE_Fader()
 {}
 
-void PGE_Fader::setFade(int speed, float target, float step)
+void PGE_Fader::setFade(int speed, double target, double step)
 {
-    fade_step = fabs(step);
+    fade_step    = fabs(step);
     target_ratio = target;
-    fadeSpeed = speed;
+    fadeSpeed    = speed;
     manual_ticks = speed;
 }
 
-void PGE_Fader::setRatio(float ratio)
+void PGE_Fader::setRatio(double ratio)
 {
-    cur_ratio=ratio;
-    target_ratio=ratio;
+    cur_ratio    = ratio;
+    target_ratio = ratio;
 }
 
-bool PGE_Fader::tickFader(float ticks)
+bool PGE_Fader::tickFader(double ticks)
 {
-    if(fadeSpeed<1) return true; //Idling animation
+    if(fadeSpeed < 1) return true; //Idling animation
     if(cur_ratio==target_ratio)
         return true;
 
-    manual_ticks-=fabs(ticks);
-        while(manual_ticks<=0.0f)
-        {
-            fadeStep();
-            manual_ticks+=fadeSpeed;
-        }
-        return (cur_ratio==target_ratio);
+    manual_ticks -= fabs(ticks);
+    while(manual_ticks<=0.0)
+    {
+        fadeStep();
+        manual_ticks += fadeSpeed;
+    }
+    return (cur_ratio==target_ratio);
 }
 
-const float &PGE_Fader::fadeRatio()
+const double &PGE_Fader::fadeRatio()
 {
     return cur_ratio;
 }
@@ -82,24 +82,24 @@ bool PGE_Fader::isFading()
 
 bool PGE_Fader::isFull()
 {
-    return ((cur_ratio>=1.0f) && (!isFading()));
+    return ((cur_ratio>=1.0) && (!isFading()));
 }
 
 void PGE_Fader::setFull()
 {
-    cur_ratio=1.0f;
-    target_ratio=1.0f;
+    cur_ratio    = 1.0;
+    target_ratio = 1.0;
 }
 
 bool PGE_Fader::isNull()
 {
-    return ((cur_ratio<=0.0f) && (!isFading()));
+    return ((cur_ratio<=0.0) && (!isFading()));
 }
 
 void PGE_Fader::setNull()
 {
-    cur_ratio    =0.0f;
-    target_ratio =0.0f;
+    cur_ratio    = 0.0;
+    target_ratio = 0.0;
 }
 
 int PGE_Fader::ticksLeft()
@@ -116,7 +116,7 @@ void PGE_Fader::fadeStep()
     else
         cur_ratio-=fade_step;
 
-    if(cur_ratio>1.0f) cur_ratio = 1.0f;
+    if(cur_ratio>1.0) cur_ratio = 1.0;
     else
-        if(cur_ratio<0.0f) cur_ratio = 0.0f;
+        if(cur_ratio<0.0) cur_ratio = 0.0;
 }
